// frontend/open-source-erp/src/Components/Inputs/Input.enum.ts
export enum InputSize {
  Sm = 'small',
  Med = 'medium',
  Lg = 'large',
}

export enum InputIconSlot {
  Pre = 'prefix',
  Suffix = 'suffix',
}

export enum InputType {
  Date = 'date',
  DateTimeLocal = 'datetime-local',
  Email = 'email',
  Number = 'number',
  Password = 'password',
  Search = 'search',
  Tel = 'tel',
  Text = 'text',
  Time = 'time',
  Url = 'url',
}